Updating for UI

+++ app/src/main/java/com/example/allgasnobrakes/models/PlayerProfile.java
@@ -1,7 +1,6 @@
 package com.example.allgasnobrakes.models;
 
 import android.util.Log;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
@@ -15,25 +14,38 @@ import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
+import java.beans.PropertyChangeListener;
+import java.beans.PropertyChangeSupport;
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.EventListener;
 import java.util.HashMap;
-import java.util.Observable;
-import java.util.Timer;
+import java.util.Locale;
 
 /**
  * Contains player profile information
  * @author zhaoyu4 zhaoyu5
- * @version 3.0
+ * @version 5.1
  */
-public class PlayerProfile extends Observable implements Serializable, EventListener {
+public class PlayerProfile implements Serializable {
+    public static final String UNIQUE_HIGHEST_RANK = "uniqueHighestRank";
+    public static final String COLLECTOR_RANK = "collectorRank";
+    private final PropertyChangeSupport pcs = new PropertyChangeSupport(this);
     private String username;
     private String email;
     private String password;
+    private int uniqueHighestRank = 0;
+    private int collectorRank = 0;
     private int displayMetric;
     private ArrayList<HashedQR> QRList = new ArrayList<>();
-    private final QRCounter profileSummary = new QRCounter(0, 0);
+    private ProfileSummary profileSummary = new ProfileSummary(0, 0);
+    private boolean test = false;
+
+    public PlayerProfile() {
+        username = "testUser";
+        test = true;
+        profileSummary = new ProfileSummary();
+        QRList.add(new HashedQR());
+    }
 
     /**
      * Constructor without password, for searching for friends account
@@ -41,7 +53,6 @@ public class PlayerProfile extends Observable implements Serializable, EventList
      * @param email The email of the account
      */
     public PlayerProfile(String username, String email) {
-        super();
         this.username = username;
         this.email = email;
     }
@@ -53,7 +64,6 @@ public class PlayerProfile extends Observable implements Serializable, EventList
      * @param password The password of the account
      */
     public PlayerProfile(String username, String email, String password) {
-        super();
         this.username = username;
         this.email = email;
         this.password = password;
@@ -69,7 +79,6 @@ public class PlayerProfile extends Observable implements Serializable, EventList
      * @param count The total number of the QR codes that account has
      */
     public PlayerProfile(String username, String email, String password, int score, int count) {
-        super();
         this.username = username;
         this.email = email;
         this.password = password;
@@ -92,16 +101,36 @@ public class PlayerProfile extends Observable implements Serializable, EventList
         return QRList;
     }
 
-    public QRCounter getProfileSummary() {
+    public ProfileSummary getProfileSummary() {
         return profileSummary;
     }
 
-    public int getTotalScore() {
-        return profileSummary.getTotalScore();
+    public HashedQR getQR(int position) {
+        return QRList.get(position);
     }
 
     public int getDisplayMetric() {
         return displayMetric;
+    }
+
+    public int getUniqueHighestRank() {
+        return uniqueHighestRank;
+    }
+
+    public int getCollectorRank() {
+        return collectorRank;
+    }
+
+    public int getNumberOfQR() {
+        return QRList.size();
+    }
+
+    public int getScore() {
+        return profileSummary.getTotalScore();
+    }
+
+    public int getCount() {
+        return profileSummary.getTotalQR();
     }
 
     public void setUsername(String username) {
@@ -116,6 +145,17 @@ public class PlayerProfile extends Observable implements Serializable, EventList
         this.displayMetric = displayMetric.intValue();
     }
 
+    public void addPropertyChangeListener(String field, PropertyChangeListener listener) {
+        pcs.addPropertyChangeListener(field, listener);
+    }
+
+    public void addScorePropertyChangeListener(String field1, String field2,
+                                               PropertyChangeListener listener1,
+                                               PropertyChangeListener listener2) {
+        profileSummary.addPropertyChangeListener(field1, listener1);
+        profileSummary.addPropertyChangeListener(field2, listener2);
+    }
+
     /**
      * Retrieves the collection of HashedQR that a player has collected and stores it locally. Also
      * notifies the view that displays this information to update itself with the latest data.
@@ -123,6 +163,9 @@ public class PlayerProfile extends Observable implements Serializable, EventList
      * @param sortOrder - the order by which to sort the QR code
      */
     public void retrieveQR(RecyclerView.Adapter QrAdapter, String sortOrder) {
+        // Notify the views to update
+        setRank();
+
         Query.Direction order;
 
         if (sortOrder.equals("Highest Score")) {
@@ -151,17 +194,14 @@ public class PlayerProfile extends Observable implements Serializable, EventList
                                         public void onComplete(@NonNull Task<DocumentSnapshot> task) {
                                             DocumentSnapshot meta = task.getResult();
                                             newQR.setComment(meta.get("Comment", String.class));
-                                            newQR.setLat(meta.get("Lat"));
-                                            newQR.setLat(meta.get("Lon"));
+                                            newQR.setLat(meta.getString("Lat"));
+                                            newQR.setLon(meta.getString("Lon"));
                                         }
                                     });
 
                             QRList.add(newQR);
                             QrAdapter.notifyDataSetChanged();
                         }
-                        // Notify the views to update
-                        setChanged();
-                        notifyObservers();
                     }
                 });
     }
@@ -171,42 +211,51 @@ public class PlayerProfile extends Observable implements Serializable, EventList
      * @param QR The QR code to be deleted
      */
     public void deleteQR(HashedQR QR) {
-        FirebaseFirestore.getInstance().document("/QR/" + QR.getHashedQR() + "/Players/" + username).delete();
-        FirebaseFirestore.getInstance().document("/QR/" + QR.getHashedQR())
-                .update("OwnedBy", FieldValue.arrayRemove(username));
+        QRList.remove(QR);
+
+        if (! test) {
+            FirebaseFirestore.getInstance().document("/QR/" + QR.getHashedQR() + "/Players/" + username).delete();
+            FirebaseFirestore.getInstance().document("/QR/" + QR.getHashedQR())
+                    .update("OwnedBy", FieldValue.arrayRemove(username))
+                    .addOnCompleteListener(new OnCompleteListener<Void>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Void> task) {
+                            updatePlayerCount(QR.getHashedQR());
+                        }
+                    });
+        }
 
         profileSummary.update(getUsername(), -1, -QR.getScore());
-        setChanged();
-        notifyObservers();
     }
 
     /**
      * Implementation of the UNDO function in case the user deleted a QR code by accident
      * @param QR The QR code to be re-added
      */
-    public void addQR(HashedQR QR) {
+    public void addQR(int position, HashedQR QR) {
+        QRList.add(position, QR);
         profileSummary.update(getUsername(), 1, QR.getScore());
-        setChanged();
-        notifyObservers();
 
-        HashMap<String, Object> meta = new HashMap<>();
+        if (! test) {
+            HashMap<String, Object> meta = new HashMap<>();
 
-        meta.put("Comment", QR.getComment());
-        meta.put("Lat", QR.getLat().toString());
-        meta.put("Lon", QR.getLon().toString());
-        FirebaseFirestore.getInstance().collection("QR").document(QR.getHashedQR())
-                .collection("Players").document(username)
-                .set(meta);
+            meta.put("Comment", QR.getComment());
+            meta.put("Lat", QR.getLat().toString());
+            meta.put("Lon", QR.getLon().toString());
+            FirebaseFirestore.getInstance().collection("QR").document(QR.getHashedQR())
+                    .collection("Players").document(username)
+                    .set(meta);
 
-        FirebaseFirestore.getInstance().collection("QR").document(QR.getHashedQR())
-                .update("OwnedBy", FieldValue.arrayUnion(username))
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        updatePlayerCount(QR.getHashedQR());
-                    }
-                });
-        Log.d("update", "2");
+            FirebaseFirestore.getInstance().collection("QR").document(QR.getHashedQR())
+                    .update("OwnedBy", FieldValue.arrayUnion(username))
+                    .addOnCompleteListener(new OnCompleteListener<Void>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Void> task) {
+                            updatePlayerCount(QR.getHashedQR());
+                        }
+                    });
+            Log.d("update", "2");
+        }
     }
 
     private void updatePlayerCount(String hash) {
@@ -217,7 +266,82 @@ public class PlayerProfile extends Observable implements Serializable, EventList
                         DocumentSnapshot qr = task.getResult();
                         int arraySize = ((ArrayList<String>) qr.get("OwnedBy")).size();
                         FirebaseFirestore.getInstance().collection("QR").document(hash)
-                                .update("PlayerCount", arraySize);
+                                .update("PlayerCount", arraySize)
+                                .addOnCompleteListener(new OnCompleteListener<Void>() {
+                                    @Override
+                                    public void onComplete(@NonNull Task<Void> task) {
+                                        Log.d("arr size", String.format(Locale.CANADA, "%d", arraySize));
+                                        setRank();
+                                    }
+                                });
+                    }
+                });
+    }
+
+    public void setRank() {
+        setUniqueHighestRank();
+        setCollectorRank();
+    }
+
+    private void setUniqueHighestRank() {
+        FirebaseFirestore.getInstance().collection("QR")
+                .whereEqualTo("PlayerCount", 1)
+                .orderBy("Score", Query.Direction.DESCENDING)
+                .limit(100)
+                .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        int rank = 1;
+                        boolean found = false;
+
+                        for (DocumentSnapshot doc: task.getResult()) {
+                            String playerName = ((ArrayList<String>) doc.get("OwnedBy")).get(0);
+                            if (username.equals(playerName)) {
+                                found = true;
+                                break;
+                            }
+                            rank++;
+                        }
+
+                        if (found) {
+                            int oldRank = uniqueHighestRank;
+                            uniqueHighestRank = rank;
+                            pcs.firePropertyChange(UNIQUE_HIGHEST_RANK, oldRank, rank);
+                        } else {
+                            uniqueHighestRank = -1;
+                            pcs.firePropertyChange(UNIQUE_HIGHEST_RANK, 0, -1);
+                        }
+                    }
+                });
+    }
+
+    private void setCollectorRank() {
+        FirebaseFirestore.getInstance().collection("Users")
+                .orderBy("Total Score", Query.Direction.DESCENDING)
+                .limit(100)
+                .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        int rank = 1;
+                        boolean found = false;
+
+                        for (DocumentSnapshot doc: task.getResult()) {
+                            String playerName = doc.getId();
+                            if(username.equals(playerName)) {
+                                found = true;
+                                break;
+                            }
+                            rank++;
+                        }
+
+                        if (found) {
+                            int oldRank = collectorRank;
+                            collectorRank = rank;
+                            pcs.firePropertyChange(COLLECTOR_RANK, oldRank, rank);
+                        } else {
+                            uniqueHighestRank = -1;
+                            pcs.firePropertyChange(COLLECTOR_RANK, 0, -1);
+                        }
                     }
                 });
     }
